# backend/internal/adapter/generated/openapi.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// Defines values for ConfirmPaymentResponsePaymentStatus.
const (
	ConfirmPaymentResponsePaymentStatusConfirmed ConfirmPaymentResponsePaymentStatus = "confirmed"
)

// Defines values for CreateOrderResponseCurrency.
const (
	CreateOrderResponseCurrencyJPY CreateOrderResponseCurrency = "JPY"
)

// Defines values for OrderCurrency.
const (
	OrderCurrencyJPY OrderCurrency = "JPY"
)

// Defines values for OrderStatus.
const (
	OrderStatusAccepted  OrderStatus = "accepted"
	OrderStatusCanceled  OrderStatus = "canceled"
	OrderStatusConfirmed OrderStatus = "confirmed"
)

// ConfirmPaymentRequest defines model for ConfirmPaymentRequest.
type ConfirmPaymentRequest struct {
	Amount         int    `json:"amount"`
	IdempotencyKey string `json:"idempotencyKey"`
	OrderId        string `json:"orderId"`
}

// ConfirmPaymentResponse defines model for ConfirmPaymentResponse.
type ConfirmPaymentResponse struct {
	OrderId       string                              `json:"orderId"`
	PaymentStatus ConfirmPaymentResponsePaymentStatus `json:"paymentStatus"`
}

// ConfirmPaymentResponsePaymentStatus defines model for ConfirmPaymentResponse.PaymentStatus.
type ConfirmPaymentResponsePaymentStatus string

// CreateOrderItem defines model for CreateOrderItem.
type CreateOrderItem struct {
	Quantity int    `json:"quantity"`
	Sku      string `json:"sku"`

	// UnitPrice クライアントが注文時点で同意した単価（サーバ価格との照合用）
	UnitPrice int `json:"unitPrice"`
}

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	CustomerId string            `json:"customerId"`
	Items      []CreateOrderItem `json:"items"`
}

// CreateOrderResponse defines model for CreateOrderResponse.
type CreateOrderResponse struct {
	// Currency 金額表示通貨（単一通貨）
	Currency *CreateOrderResponseCurrency `json:"currency,omitempty"`
	OrderId  string                       `json:"orderId"`
	Status   OrderStatus                  `json:"status"`

	// TotalAmount 注文合計金額（サーバ算出、最小通貨単位）
	TotalAmount *int `json:"totalAmount,omitempty"`
}

// CreateOrderResponseCurrency 金額表示通貨（単一通貨）
type CreateOrderResponseCurrency string

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// Order defines model for Order.
type Order struct {
	// Currency 金額表示通貨（単一通貨）
	Currency   *OrderCurrency `json:"currency,omitempty"`
	CustomerId string         `json:"customerId"`
	Id         string         `json:"id"`
	Items      []OrderItem    `json:"items"`
	Status     OrderStatus    `json:"status"`

	// TotalAmount 注文合計金額（サーバ算出、最小通貨単位）
	TotalAmount *int `json:"totalAmount,omitempty"`
}

// OrderCurrency 金額表示通貨（単一通貨）
type OrderCurrency string

// OrderItem defines model for OrderItem.
type OrderItem struct {
	// LineAmount 明細金額（unitPrice * quantity）
	LineAmount *int   `json:"lineAmount,omitempty"`
	Quantity   int    `json:"quantity"`
	Sku        string `json:"sku"`

	// UnitPrice 商品単価（サーバ側価格情報を決定元に設定）
	UnitPrice *int `json:"unitPrice,omitempty"`
}

// OrderStatus defines model for OrderStatus.
type OrderStatus string

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody = CreateOrderRequest

// ConfirmPaymentJSONRequestBody defines body for ConfirmPayment for application/json ContentType.
type ConfirmPaymentJSONRequestBody = ConfirmPaymentRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create order (reserve inventory)
	// (POST /orders)
	CreateOrder(c *gin.Context)
	// Get order by id
	// (GET /orders/{orderId})
	GetOrder(c *gin.Context, orderId string)
	// Cancel order (release inventory)
	// (POST /orders/{orderId}/cancel)
	CancelOrder(c *gin.Context, orderId string)
	// Confirm payment (idempotent)
	// (POST /payments/confirm)
	ConfirmPayment(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// CreateOrder operation middleware
func (siw *ServerInterfaceWrapper) CreateOrder(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateOrder(c)
}

// GetOrder operation middleware
func (siw *ServerInterfaceWrapper) GetOrder(c *gin.Context) {

	var err error

	// ------------- Path parameter "orderId" -------------
	var orderId string

	err = runtime.BindStyledParameterWithOptions("simple", "orderId", c.Param("orderId"), &orderId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter orderId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetOrder(c, orderId)
}

// CancelOrder operation middleware
func (siw *ServerInterfaceWrapper) CancelOrder(c *gin.Context) {

	var err error

	// ------------- Path parameter "orderId" -------------
	var orderId string

	err = runtime.BindStyledParameterWithOptions("simple", "orderId", c.Param("orderId"), &orderId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter orderId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CancelOrder(c, orderId)
}

// ConfirmPayment operation middleware
func (siw *ServerInterfaceWrapper) ConfirmPayment(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ConfirmPayment(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/orders", wrapper.CreateOrder)
	router.GET(options.BaseURL+"/orders/:orderId", wrapper.GetOrder)
	router.POST(options.BaseURL+"/orders/:orderId/cancel", wrapper.CancelOrder)
	router.POST(options.BaseURL+"/payments/confirm", wrapper.ConfirmPayment)
}