# contracts/openapi.yaml
openapi: 3.0.3
# 境界の単一参照点（契約の一次情報）
# ここ以外に境界仕様を書かない前提で、生成/差分/テストの基準にする
info:
  title: Order Inventory API
  version: 0.1.0
  description: |
    在庫は標準モデル（OnHand / Reserved / Available）で扱う。
    Available は OnHand - Reserved で定義する。
    注文作成で Reserved を増加（Reserve）、キャンセルで Reserved を減少（Release）する。
    決済確定では在庫を減算しない（OnHand / Reserved は不変）。
    出荷機能は本契約の対象外であり、出荷確定での OnHand 減算は未導入。
    価格モデルはサーバ側価格情報を決定元とし、クライアント入力価格は採用しない。
    注文金額はサーバ側で算出し、注文応答に金額項目として返却する。
servers:
  - url: http://localhost:8080
paths:
  /orders:
    post:
      summary: Create order (reserve inventory)
      operationId: createOrder
      description: |
        価格モデル前提:
        - リクエストに単価は含めない
        - `unitPrice` / `lineAmount` / `totalAmount` はサーバ算出値を返す
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Order accepted
          # 200の意味を固定する（accepted）
          # 200が返る＝注文が受理され、以降の参照で確認可能である前提
          # 在庫前提:
          # - 注文作成時に Reserve を実行し、Reserved を増加させる
          # - OnHand は変更しない
          # 価格前提:
          # - 応答の金額項目はサーバ算出値を返す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Price conflict (quoted unit price mismatches server-side product price)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{orderId}:
    get:
      summary: Get order by id
      operationId: getOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          # 存在しないIDは404とする前提
          # 200 + 空レスポンスではなく、明示的にNot Foundを返す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{orderId}/cancel:
    post:
      summary: Cancel order (release inventory)
      operationId: cancelOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order canceled
          # 在庫前提:
          # - キャンセル時に Release を実行し、Reserved を減少させる
          # - OnHand は変更しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/confirm:
    post:
      summary: Confirm payment (idempotent)
      operationId: confirmPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmPaymentRequest'
      responses:
        '200':
          description: Payment confirmed
          # 冪等性の固定は境界テストで担保する
          # 同一 idempotencyKey での二重課金が起きないことが条件
          # 在庫前提:
          # - 決済確定では在庫を減算しない
          # - OnHand / Reserved は不変
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmPaymentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Amount conflict (mismatch or idempotency key reused with different amount)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
      properties:
        customerId:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItem'
    CreateOrderItem:
      type: object
      required:
        - sku
        - quantity
        - unitPrice
      properties:
        sku:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: integer
          minimum: 0
          description: クライアントが注文時点で同意した単価（サーバ価格との照合用）
    CreateOrderResponse:
      type: object
      required:
        - orderId
        - status
      properties:
        orderId:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        currency:
          type: string
          enum:
            - JPY
          description: 金額表示通貨（単一通貨）
        totalAmount:
          type: integer
          minimum: 0
          description: 注文合計金額（サーバ算出、最小通貨単位）
    ConfirmPaymentRequest:
      type: object
      required:
        - orderId
        - amount
        - idempotencyKey
      properties:
        orderId:
          type: string
        amount:
          type: integer
        idempotencyKey:
          type: string
    ConfirmPaymentResponse:
      type: object
      required:
        - orderId
        - paymentStatus
      properties:
        orderId:
          type: string
        paymentStatus:
          type: string
          enum:
            - confirmed
    Order:
      type: object
      required:
        - id
        - customerId
        - status
        - items
      properties:
        id:
          type: string
        customerId:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        currency:
          type: string
          enum:
            - JPY
          description: 金額表示通貨（単一通貨）
        totalAmount:
          type: integer
          minimum: 0
          description: 注文合計金額（サーバ算出、最小通貨単位）
    OrderItem:
      type: object
      required:
        - sku
        - quantity
      properties:
        sku:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: integer
          minimum: 0
          description: 商品単価（サーバ側価格情報を決定元に設定）
        lineAmount:
          type: integer
          minimum: 0
          description: 明細金額（unitPrice * quantity）
    OrderStatus:
      type: string
      enum:
        - accepted
        - confirmed
        - canceled
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    InventoryBalance:
      type: object
      description: |
        標準在庫モデルの在庫表現。
        Available は OnHand - Reserved で定義する。
      required:
        - onHand
        - reserved
        - available
      properties:
        onHand:
          type: integer
          minimum: 0
          description: 実在庫（出荷確定で減算対象）
        reserved:
          type: integer
          minimum: 0
          description: 引当済み在庫（注文作成で増加、キャンセル/期限切れで減少）
        available:
          type: integer
          minimum: 0
          description: 販売可能在庫（onHand - reserved）